Migrate microservice to alpine linux
Our fedora21 stack is outdated, unsupported and huge. This image size
bigger than 800M and that now that we use the public registry is a
problem.

With this commit, we migrate to alpine linux which allow us to have
images around 100~150M plus many other benefits

The commit-stage installs whatever is necessary to pass the test and
other things that happens within grunt taks.